+++ packages/core/src/uiengine.js
@@ -152,7 +152,7 @@ const startServer = (state, opts) => {
         // whole iframe host in case just the iframe content changes.
         // as the browser-sync client scripts loads asynchronously, we may need to
         // retrigger this function a few times.
-        return match + `\n<!-- UIengine: inject start -->\n${snippet}\n<script>
+        return match + `\n\n<!-- UIengine: inject start -->\n${snippet}\n<script>
         let retries = 0;
         function setupSocket () {
           const socket = window.___browserSync___ && window.___browserSync___.socket;
@@ -172,7 +172,7 @@ const startServer = (state, opts) => {
           }
         }
         setupSocket();
-        </script>\n<!-- UIengine: inject end -->`
+        </script>\n<!-- UIengine: inject end -->\n\n`
       }
     }
   }